+++ engine/src/main/java/dev/gamekit/ui/widgets/NinePatch.java
@@ -49,7 +49,10 @@ public class NinePatch extends SingleChildParent {
       constraints.constrainHeight(intrinsicBounds.height)
     );
 
-    child.computedBounds.setPosition(0, 0);
+    child.computedBounds.setPosition(
+      computedBounds.width / 2 - child.computedBounds.width / 2,
+      computedBounds.height / 2 - child.computedBounds.height / 2
+    );
   }
 
   @Override