+++ engine/src/main/java/dev/gamekit/ui/widgets/Padding.java
@@ -47,7 +47,10 @@ public class Padding extends SingleChildParent {
       );
     }
 
-    child.computedBounds.setPosition(padding.left, padding.top);
+    child.computedBounds.setPosition(
+      computedBounds.width / 2 - child.computedBounds.width / 2,
+      computedBounds.height / 2 - child.computedBounds.height / 2
+    );
   }
 
   @Override